user routing

+++ Larvel/app/Http/Controllers/eduprogramController.php
@@ -6,6 +6,7 @@ use App\Models\Eduprogram;
 use Illuminate\Support\Facades\DB;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class eduprogramController extends Controller
 {
@@ -18,10 +19,14 @@ class eduprogramController extends Controller
     {
         // return "asldfkja";
         // return $request->method();
-        $programs =  Eduprogram::orderBy('id', 'desc')->paginate(10);
-        $path = '/educationalProgramList';
-        // return $programs;
-        return view('admin.pdc-list-all-educational-program', compact('programs', 'path'));
+        if (Gate::allows(ability: 'is-admin')) {
+            $programs =  Eduprogram::orderBy('id', 'desc')->paginate(10);
+            $path = '/educationalProgramList';
+            // return $programs;
+            return view('admin.pdc-list-all-educational-program', compact('programs', 'path'));
+        } else {
+            dd('you need to be admin');
+        }
     }
 
     /**
@@ -44,26 +49,120 @@ class eduprogramController extends Controller
     {
         //
         // return $request->path();
-        if ($request->path() === 'admin/searchEducationalProgram') {
-            // $request->validate([
-            //     'search_type' => 'bail|required|string|in:year,department,faculty,university,teacher_last_name,teacher_name,teacher_name,type,topic',
-            //     'search_content' => 'bail|required',
-            // ]);
-            if ($request->search_type === null || $request->search_content === null) {
-                return redirect('admin/educationalProgramList');
-            }
-            $path = '/admin/educationalProgramList';
-            $programs =  DB::table('eduprograms')->where($request->search_type, $request->search_content)->paginate(10);
-            // return count($programs);
-            if (count($programs) === 0) {
-                return back()->with('warn_search', 'یاد پروګرام په سیسټم کي ونه موندل سو!');
+        if (Gate::allows(ability: 'is-admin')) {
+            if ($request->path() === 'admin/searchEducationalProgram') {
+                // $request->validate([
+                //     'search_type' => 'bail|required|string|in:year,department,faculty,university,teacher_last_name,teacher_name,teacher_name,type,topic',
+                //     'search_content' => 'bail|required',
+                // ]);
+                if ($request->search_type === null || $request->search_content === null) {
+                    return redirect('admin/educationalProgramList');
+                }
+                $path = '/admin/educationalProgramList';
+                $programs =  DB::table('eduprograms')->where($request->search_type, $request->search_content)->paginate(10);
+                // return count($programs);
+                if (count($programs) === 0) {
+                    return back()->with('warn_search', 'یاد پروګرام په سیسټم کي ونه موندل سو!');
+                } else {
+                    // return "asdf";
+                    // redirect('admin/educafdfdftionalProgramList')->with('success_search', 'لاندي ستاسي پلټل سوی پروګرام دی!');
+                    return view('admin.pdc-list-all-educational-program', compact('programs', 'path'))->with('success_search', 'لاندي ستاسي پلټل سوی پروګرام دی!');
+                }
+            } else if ($request->path() === 'admin/educationalProgramList') {
+                // return $request->date;
+                $request->validate([
+                    'topic' => 'bail|required|string|max:100',
+                    'type' => 'bail|required|string|in:تقرر,ارتقاء,علمي ترفېع',
+                    'teacher_name' => 'bail|required|string|max:30',
+                    'father_name' => 'bail|required|string|max:30',
+                    'teacher_last_name' => 'bail|required|string|max:30',
+                    'university' => 'bail|nullable|string|in:کندهار پوهنتون',
+                    'faculty' => 'bail|required|string|in:کمپیوټر ساینس,انجنیري,حقوق,اداره ئې عامه,ژورنالیزم,اقتصاد,زراعت,شرعیات,ادبیات,ساینس,تعلیم او تربیه',
+                    'department' => 'bail|required|string|max:30',
+                    'current_educational_position' => 'bail|required|string|in:پوهاند,پوهنمل,پوهنیار,پوهیالی',
+                    'achieving_educational_position' => 'bail|required|string|in:پوهاند,پوهنمل,پوهنیار,پوهیالی',
+                    'participant_amount' => 'bail|required|integer|between:1,200',
+                    'date' => 'bail|required|date',
+                    // 'month' => 'bail|required|integer|between:1,12',
+                    // 'start_day' => 'bail|required|integer|between:1,31',
+                    // 'start_time' => 'bail|required|date_format:H:i',
+                    'campus_name' => 'bail|required|string|max:30',
+                    'block_name' => 'bail|required|string|max:30',
+                    'block_number' => 'bail|required|integer|between:1,30',
+                    'room_number' => 'bail|required|integer|between:1,30',
+                ]);
+                $program = new Eduprogram;
+                $program->topic = $request->topic;
+                $program->type = $request->type;
+                $program->teacher_name = $request->teacher_name;
+                $program->father_name = $request->father_name;
+                $program->teacher_last_name = $request->teacher_last_name;
+                $program->university = $request->university;
+                $program->faculty = $request->faculty;
+                $program->department = $request->department;
+                $program->current_educational_position = $request->current_educational_position;
+                $program->achieving_educational_position = $request->achieving_educational_position;
+                $program->participant_amount = $request->participant_amount;
+                $program->date = $request->date;
+                $program->campus_name = $request->campus_name;
+                $program->block_name = $request->block_name;
+                $program->block_number = $request->block_number;
+                $program->room_number = $request->room_number;
+                $program->save();
+
+                return back()->with('success', 'یاد پروګرام په کامیابۍ سره سیسټم اضافه کړل سو!');
             } else {
-                // return "asdf";
-                // redirect('admin/educafdfdftionalProgramList')->with('success_search', 'لاندي ستاسي پلټل سوی پروګرام دی!');
-                return view('admin.pdc-list-all-educational-program', compact('programs', 'path'))->with('success_search', 'لاندي ستاسي پلټل سوی پروګرام دی!');
+                return "not found path!!!";
             }
-        } else if ($request->path() === 'admin/educationalProgramList') {
-            // return $request->date;
+        } else {
+            dd('you need to be admin');
+        }
+    }
+
+    /**
+     * Display the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function show($id)
+    {
+        //
+        if (Gate::allows(ability: 'is-admin')) {
+            $program = Eduprogram::find($id);
+            return view('admin.pdc-educational-program-info', compact('program'));
+        } else {
+            dd('you need to be admin');
+        }
+    }
+
+    /**
+     * Show the form for editing the specified resource.
+     *
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function edit($id)
+    {
+        if (Gate::allows(ability: 'is-admin')) {
+            $program = Eduprogram::find($id);
+            return view('admin.pdc-edit-educational-program', compact('program'));
+        } else {
+            dd('you need to be admin');
+        }
+    }
+
+    /**
+     * Update the specified resource in storage.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     * @param  int  $id
+     * @return \Illuminate\Http\Response
+     */
+    public function update(Request $request, $id)
+    {
+        // return $request->path();
+        if (Gate::allows(ability: 'is-admin')) {
             $request->validate([
                 'topic' => 'bail|required|string|max:100',
                 'type' => 'bail|required|string|in:تقرر,ارتقاء,علمي ترفېع',
@@ -77,15 +176,13 @@ class eduprogramController extends Controller
                 'achieving_educational_position' => 'bail|required|string|in:پوهاند,پوهنمل,پوهنیار,پوهیالی',
                 'participant_amount' => 'bail|required|integer|between:1,200',
                 'date' => 'bail|required|date',
-                // 'month' => 'bail|required|integer|between:1,12',
-                // 'start_day' => 'bail|required|integer|between:1,31',
-                // 'start_time' => 'bail|required|date_format:H:i',
                 'campus_name' => 'bail|required|string|max:30',
                 'block_name' => 'bail|required|string|max:30',
                 'block_number' => 'bail|required|integer|between:1,30',
                 'room_number' => 'bail|required|integer|between:1,30',
             ]);
-            $program = new Eduprogram;
+            // return "sdfsdfs";
+            $program = Eduprogram::find($id);
             $program->topic = $request->topic;
             $program->type = $request->type;
             $program->teacher_name = $request->teacher_name;
@@ -103,88 +200,10 @@ class eduprogramController extends Controller
             $program->block_number = $request->block_number;
             $program->room_number = $request->room_number;
             $program->save();
-
-            return back()->with('success', 'یاد پروګرام په کامیابۍ سره سیسټم اضافه کړل سو!');
+            return redirect('admin/educationalProgramList')->with('success', 'د یاد پروګرام معلومات په سیسټم کي په کامیابۍ سره تغیر ورکړل سو!');
         } else {
-            return "not found path!!!";
+            dd('you need to be admin');
         }
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-        $program = Eduprogram::find($id);
-        return view('admin.pdc-educational-program-info', compact('program'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function edit($id)
-    {
-
-        $program = Eduprogram::find($id);
-        return view('admin.pdc-edit-educational-program', compact('program'));
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, $id)
-    {
-        // return $request->path();
-
-        $request->validate([
-            'topic' => 'bail|required|string|max:100',
-            'type' => 'bail|required|string|in:تقرر,ارتقاء,علمي ترفېع',
-            'teacher_name' => 'bail|required|string|max:30',
-            'father_name' => 'bail|required|string|max:30',
-            'teacher_last_name' => 'bail|required|string|max:30',
-            'university' => 'bail|nullable|string|in:کندهار پوهنتون',
-            'faculty' => 'bail|required|string|in:کمپیوټر ساینس,انجنیري,حقوق,اداره ئې عامه,ژورنالیزم,اقتصاد,زراعت,شرعیات,ادبیات,ساینس,تعلیم او تربیه',
-            'department' => 'bail|required|string|max:30',
-            'current_educational_position' => 'bail|required|string|in:پوهاند,پوهنمل,پوهنیار,پوهیالی',
-            'achieving_educational_position' => 'bail|required|string|in:پوهاند,پوهنمل,پوهنیار,پوهیالی',
-            'participant_amount' => 'bail|required|integer|between:1,200',
-            'date' => 'bail|required|date',
-            'campus_name' => 'bail|required|string|max:30',
-            'block_name' => 'bail|required|string|max:30',
-            'block_number' => 'bail|required|integer|between:1,30',
-            'room_number' => 'bail|required|integer|between:1,30',
-        ]);
-        // return "sdfsdfs";
-        $program = Eduprogram::find($id);
-        $program->topic = $request->topic;
-        $program->type = $request->type;
-        $program->teacher_name = $request->teacher_name;
-        $program->father_name = $request->father_name;
-        $program->teacher_last_name = $request->teacher_last_name;
-        $program->university = $request->university;
-        $program->faculty = $request->faculty;
-        $program->department = $request->department;
-        $program->current_educational_position = $request->current_educational_position;
-        $program->achieving_educational_position = $request->achieving_educational_position;
-        $program->participant_amount = $request->participant_amount;
-        $program->date = $request->date;
-        $program->campus_name = $request->campus_name;
-        $program->block_name = $request->block_name;
-        $program->block_number = $request->block_number;
-        $program->room_number = $request->room_number;
-        $program->save();
-        return redirect('admin/educationalProgramList')->with('success', 'د یاد پروګرام معلومات په سیسټم کي په کامیابۍ سره تغیر ورکړل سو!');
     }
 
     /**
@@ -195,12 +214,16 @@ class eduprogramController extends Controller
      */
     public function destroy($id)
     {
-        if (count(DB::table('eduprograms')->where('id', $id)->get()) === 1) {
-            $program = Eduprogram::find($id);
-            $program->delete();
-            return redirect('admin/educationalProgramList')->with('success', 'د یاد پروګرام معلومات له سیسټم څخه په کامیابۍ سره له منځه لاړ!');
-        } elseif (count(DB::table('eduprograms')->where('id', $id)->get()) === 0) {
-            return redirect('admin/educationalProgramList')->with('warn', 'یاد پروګرام په سیسټم کي د له منځه وړلو لپاره نسو پیدا!');
+        if (Gate::allows(ability: 'is-admin')) {
+            if (count(DB::table('eduprograms')->where('id', $id)->get()) === 1) {
+                $program = Eduprogram::find($id);
+                $program->delete();
+                return redirect('admin/educationalProgramList')->with('success', 'د یاد پروګرام معلومات له سیسټم څخه په کامیابۍ سره له منځه لاړ!');
+            } elseif (count(DB::table('eduprograms')->where('id', $id)->get()) === 0) {
+                return redirect('admin/educationalProgramList')->with('warn', 'یاد پروګرام په سیسټم کي د له منځه وړلو لپاره نسو پیدا!');
+            }
+        } else {
+            dd('you need to be admin');
         }
