+++ Larvel/app/Http/Controllers/materialController.php
@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\DB;
 use App\Models\Program;
 use Illuminate\Http\Request;
 use File;
+use Illuminate\Support\Facades\Gate;
 
 use Illuminate\Support\Facades\Storage;
 // use Storage;
@@ -43,37 +44,43 @@ class materialController extends Controller
     public function store(Request $request)
     {
         // return "sljdkfj";
-
-        if ($request->path() == 'admin/storeMaterials') {
-            $index = 0;
-            // return "sldkfjsdf";
-            // return $request->file('materials');
-            foreach ($request->file('materials') as $material) {
-                if (!empty($request->file_name[$index]) && !empty($request->file_type[$index])) {
-                    // return "slakdj";
-                    $fileName = time() . '.' . $material->extension();
-                    $kb = $material->getSize() / 1024;
-                    $mb = $kb / 1024;
-                    $gb = $mb / 1024;
-                    $material->storeAs('public/programFiles', $fileName);
-                    sleep(1);
-                    $fileSave = new Material;
-                    $fileSave->name = $request->file_name[$index];
-                    $fileSave->type = $request->file_type[$index];
-                    $fileSave->path = $fileName;
-                    $fileSave->size =  round($mb, 2);
-                    $fileSave->extension =  $material->extension();
-                    $fileSave->program_id = $request->program_id;
-                    $fileSave->save();
-                    $index++;
-                } elseif (empty($request->file_name[$index])) {
-                    return back()->with('warn', "د فایل نوم باید وجود ولري!");
-                } elseif (empty($request->file_type[$index])) {
-                    return back()->with('warn', "د فایل ډول باید وجود ولري!");
+        if (Gate::allows(ability: 'is-admin')) {
+            if ($request->path() == 'admin/storeMaterials') {
+                $index = 0;
+                // return "sldkfjsdf";
+                // return $request->file('materials');
+                foreach ($request->file('materials') as $material) {
+                    if (!empty($request->file_name[$index]) && !empty($request->file_type[$index])) {
+                        // return "slakdj";
+                        $fileName = time() . '.' . $material->extension();
+                        $kb = $material->getSize() / 1024;
+                        $mb = $kb / 1024;
+                        $gb = $mb / 1024;
+                        $material->storeAs('public/programFiles', $fileName);
+                        sleep(1);
+                        $fileSave = new Material;
+                        $fileSave->name = $request->file_name[$index];
+                        $fileSave->type = $request->file_type[$index];
+                        $fileSave->path = $fileName;
+                        $fileSave->size =  round($mb, 2);
+                        $fileSave->extension =  $material->extension();
+                        $fileSave->program_id = $request->program_id;
+                        $fileSave->save();
+                        $index++;
+                    } elseif (empty($request->file_name[$index])) {
+                        return back()->with('warn', "د فایل نوم باید وجود ولري!");
+                    } elseif (empty($request->file_type[$index])) {
+                        return back()->with('warn', "د فایل ډول باید وجود ولري!");
+                    }
                 }
+                return redirect('admin/pdcProgramInfo/' . $request->program_id)->with('program_materials_added', "پروګرام اړونده فایلونه په کامیابۍ سره سیسټم ته داخل کړل سوه!");
             }
-            return redirect('admin/pdcProgramInfo/' . $request->program_id)->with('program_materials_added', "پروګرام اړونده فایلونه په کامیابۍ سره سیسټم ته داخل کړل سوه!");
-        } elseif ($request->path() == 'user/storeMaterials') {
+        } else {
+            dd('you need to be admin');
+        }
+
+
+        if ($request->path() == 'user/storeMaterials') {
             // return "i am store mat";
             $request->validate([
                 'materials' => 'required',
@@ -123,37 +130,43 @@ class materialController extends Controller
     public function show(Request $request, $id)
     {
         //
-        if ($request->path() === 'admin/storeMaterials/' . $id) {
-            $program_id = $id;
-            return view('admin.pdc-add-programs-materials', compact('program_id'));
-        } elseif ($request->path() === 'admin/materials/' . $id) {
-            $program_id = $id;
-            // return $program_id;
-            $programMaterials = Program::with('getMaterials')->find($id);
-            $programMaterialsCheck = DB::table('materials')->where('program_id', $id)->get();
+        if (Gate::allows(ability: 'is-admin')) {
+            if ($request->path() === 'admin/storeMaterials/' . $id) {
+                $program_id = $id;
+                return view('admin.pdc-add-programs-materials', compact('program_id'));
+            } elseif ($request->path() === 'admin/materials/' . $id) {
+                $program_id = $id;
+                // return $program_id;
+                $programMaterials = Program::with('getMaterials')->find($id);
+                $programMaterialsCheck = DB::table('materials')->where('program_id', $id)->get();
 
 
-            if (count($programMaterialsCheck) !== 0) {
-                return view('admin.files-download', compact('programMaterials', 'program_id'));
-            } else {
-                return back()->with('warn', 'د یاد پروګرام لپاره تر اوسه فایلونه ندي اضافه سوي سیسټم کي !');
+                if (count($programMaterialsCheck) !== 0) {
+                    return view('admin.files-download', compact('programMaterials', 'program_id'));
+                } else {
+                    return back()->with('warn', 'د یاد پروګرام لپاره تر اوسه فایلونه ندي اضافه سوي سیسټم کي !');
+                }
+            } elseif ($request->path() === 'downloadMaterial/' . $id) {
+                // return "hi";
+                return Storage::download('public/programFiles/' . $id);
+                // return Storage::download('public/programFiles/'.$id, $name, $headers);
+            } elseif ($request->path() === 'admin/downloadMaterial/' . $id) {
+                return Storage::download('public/programFiles/' . $id);
+                return $id;
+            } elseif ($request->path() === 'admin/viewMaterial/' . $id) {
+                // return Storage::download('public/programFiles/' . $id);
+                $path = 'storage/programFiles/' . $id;
+                return view('admin.viewFile', compact('path'));
+
+                // return $path;
             }
-        } elseif ($request->path() === 'user/facilitatorMaterials/' . $id) {
+        } else {
+            // dd('you need to be admddddin');
+        }
+
+        if ($request->path() === 'user/facilitatorMaterials/' . $id) {
             $program_id = $id;
             return view('users.pdc-add-programs-materials', compact('program_id'));
-        } elseif ($request->path() === 'downloadMaterial/' . $id) {
-            // return "hi";
-            return Storage::download('public/programFiles/' . $id);
-            // return Storage::download('public/programFiles/'.$id, $name, $headers);
-        } elseif ($request->path() === 'admin/downloadMaterial/' . $id) {
-            return Storage::download('public/programFiles/' . $id);
-            return $id;
-        } elseif ($request->path() === 'admin/viewMaterial/' . $id) {
-            // return Storage::download('public/programFiles/' . $id);
-            $path = 'storage/programFiles/' . $id;
-            return view('admin.viewFile', compact('path'));
-
-            // return $path;
         }
     }
 
@@ -194,19 +207,23 @@ class materialController extends Controller
         // return redirect('materials/'.$request->program_id);
         // return $request->program_id;
         // return $id;
-        Storage::delete('public/programFiles/' . $id);
-        // return "del";
-        $delete = DB::table('materials')->where('materials.path', $id)->get();
-        if ($delete[0]->program_id == $request->program_id) {
-            DB::table('materials')->where('materials.path', $id)->delete();
+        if (Gate::allows(ability: 'is-admin')) {
+            Storage::delete('public/programFiles/' . $id);
+            // return "del";
+            $delete = DB::table('materials')->where('materials.path', $id)->get();
+            if ($delete[0]->program_id == $request->program_id) {
+                DB::table('materials')->where('materials.path', $id)->delete();
+            } else {
+                return back()->with('warn', " یاد فایل چي تاسي غواړی  له منځه یوسی پدې پروګرام پوري اړه نلري!");
+            }
+            $check = DB::table('materials')->where('program_id', $request->program_id)->get();
+            if (count($check) === 0) {
+                return redirect('admin/pdcProgramInfo/' . $request->program_id)->with('success', "د یاد پروګرام ټوله فایلونه له سیسټم څخه له منځه ولاړ!");;
+            } else {
+                return back()->with('success', "یاد فایل له سیسټم څخه له منځه ولاړ!");
+            }
         } else {
-            return back()->with('warn', " یاد فایل چي تاسي غواړی  له منځه یوسی پدې پروګرام پوري اړه نلري!");
-        }
-        $check = DB::table('materials')->where('program_id', $request->program_id)->get();
-        if (count($check) === 0) {
-            return redirect('admin/pdcProgramInfo/' . $request->program_id)->with('success', "د یاد پروګرام ټوله فایلونه له سیسټم څخه له منځه ولاړ!");;
-        } else {
-            return back()->with('success', "یاد فایل له سیسټم څخه له منځه ولاړ!");
+            dd('you need to be admin');
         }
     }
 }
