+++ Larvel/app/Http/Controllers/photoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Photo;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class photoController extends Controller
 {
@@ -38,24 +39,28 @@ class photoController extends Controller
     {
         // return count($request->image);
         // return $photos = $request->file('image');
-        $request->validate([
-            'photo' => 'required',
-            'photo.*' => 'mimes:jpeg,jpg,png,gif,csv|max:10240',
-        ]);
-        // $request->validate([
-        //     'photo.*' => 'bail|required|image|mimes:jpeg,png,jpg,gif,svg|max:10240',
-        // ]);
+        if (Gate::allows(ability: 'is-admin')) {
+            $request->validate([
+                'photo' => 'required',
+                'photo.*' => 'mimes:jpeg,jpg,png,gif,csv|max:10240',
+            ]);
+            // $request->validate([
+            //     'photo.*' => 'bail|required|image|mimes:jpeg,png,jpg,gif,svg|max:10240',
+            // ]);
 
-        foreach ($request->file('photo') as $photo) {
-            $imageName = time() . '.' . $photo->extension();
-            $photo->storeAs('public/images', $imageName);
-            sleep(1);
-            $imageSave = new Photo;
-            $imageSave->program_id = $request->program_id;
-            $imageSave->path = $imageName;
-            $imageSave->save();
+            foreach ($request->file('photo') as $photo) {
+                $imageName = time() . '.' . $photo->extension();
+                $photo->storeAs('public/images', $imageName);
+                sleep(1);
+                $imageSave = new Photo;
+                $imageSave->program_id = $request->program_id;
+                $imageSave->path = $imageName;
+                $imageSave->save();
+            }
+            return redirect('admin/pdcProgramInfo/' . $request->program_id)->with('program_part_added', "پروګرام اړونده تفرقې په کامیابۍ سره سیسټم ته داخل کړل سوه!");
+        } else {
+            dd('you need to be admin');
         }
-        return redirect('admin/pdcProgramInfo/' . $request->program_id)->with('program_part_added', "پروګرام اړونده تفرقې په کامیابۍ سره سیسټم ته داخل کړل سوه!");
     }
 
     /**
@@ -67,11 +72,15 @@ class photoController extends Controller
     public function show(Request $request, $id)
     {
         //
-        if ($request->path() === 'admin/pdcProgramPhoto/' . $id) {
-            $programID = $id;
-            return view('admin.pdc-program-photo', compact('programID'));
-        } elseif ($request->path() === 'admin/downloadPhoto/' . $id) {
-            return Storage::download('public/images/' . $id);
+        if (Gate::allows(ability: 'is-admin')) {
+            if ($request->path() === 'admin/pdcProgramPhoto/' . $id) {
+                $programID = $id;
+                return view('admin.pdc-program-photo', compact('programID'));
+            } elseif ($request->path() === 'admin/downloadPhoto/' . $id) {
+                return Storage::download('public/images/' . $id);
+            }
+        } else {
+            dd('you need to be admin');
         }
 
         // return "i am photo!";
@@ -111,9 +120,13 @@ class photoController extends Controller
         //
         // return "hahahahmmmmmmmmma";
         // return $id;
-        Storage::delete('public/images/' . $id);
-        Photo::where('path', $id)->delete();
-        // $deletematerial->delete();
-        return redirect('admin/pdcProgramInfo/' . $request->program_id);
+        if (Gate::allows(ability: 'is-admin')) {
+            Storage::delete('public/images/' . $id);
+            Photo::where('path', $id)->delete();
+            // $deletematerial->delete();
+            return redirect('admin/pdcProgramInfo/' . $request->program_id);
+        } else {
+            dd('you need to be admin');
+        }
     }
 }
